Added changes for NDD PR
